requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
idna==3.3
phonenumbers==8.12.45
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
pyotp==2.6.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
twilio==7.7.1
tzdata==2021.5
urllib3==1.26.9
